feat(wallet): extract connection logic to hook
